Made SkillPoints usabe
Player now spends his SP to level up a skill and retreives them when
leveling it down. Also changed the way SP is stored. It no longer uses a
hashMap. Instead, it is gotten and set directly to the players.yml. Also
added feedback message when a command is done wrong or can't be used.